pip-requirements.txt
python-telegram-bot==13.4.1
pyowm==3.2.0